mod one hot
