+++ tfjs-core/src/gradients/OneHot_grad.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2020 Google Inc. All Rights Reserved.
+ * Copyright 2019 Google Inc. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -15,6 +15,16 @@
  * =============================================================================
  */
 
-import './squared_difference';
-import './broadcast_to';
-import './one_hot';
+import {OneHot} from '../kernel_names';
+import {GradConfig} from '../kernel_registry';
+import {zeros} from '../ops/tensor_ops';
+import {Tensor} from '../tensor';
+
+export const oneHotGradConfig: GradConfig = {
+  kernelName: OneHot,
+  inputsToSave: ['indices'],
+  gradFunc: (dy: Tensor, saved: Tensor[]) => {
+    const indices = saved[0];
+    return {indices: () => zeros(indices.shape, 'float32')};
+  }
+};
