+++ tfjs-core/src/tests.ts
@@ -74,6 +74,7 @@ import './ops/lstm_test';
 import './ops/matmul_test';
 import './ops/moving_average_test';
 import './ops/multinomial_test';
+import './ops/one_hot_test';
 import './ops/operation_test';
 import './ops/pad_test';
 import './ops/pool_test';
